Fixed Session Role Page

+++ admin/deleteBuku.php
@@ -2,6 +2,10 @@
     require '../dbController.php';
 
     $id = $_GET["id"];
+    if (!isset($_SESSION["loggedin"]) || $_SESSION["loggedin"] !== true || isset($_SESSION['Admin'])) {
+      header("Location: ../login.php");
+      exit;
+    }
 
     if (hapusBuku($id) > 0) {
       echo "
